+++ ExpandWorldPrefabs/service/data/ZdoHelper.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using ExpandWorld.Prefab;
 using Service;
 using UnityEngine;
 
@@ -103,4 +102,41 @@ public static class ZdoHelper
     return key.GetStableHashCode();
   }
   public static string ReverseHash(int hash) => ReverseHashCache.TryGetValue(hash, out var k) ? k : hash.ToString(CultureInfo.InvariantCulture);
+
+  private static readonly int InventoryWidthHash = Hash("Container.m_width");
+  private static readonly int InventoryHeightHash = Hash("Container.m_height");
+  public static Vector2i GetInventorySize(DataEntry entry, Parameters parameters, ZDO zdo)
+  {
+    // Width and height can come from data entry, existing ZDO fields or directly from the prefab.
+    int width = 0;
+    int height = 0;
+
+    if (entry.Ints?.TryGetValue(InventoryWidthHash, out var w) == true)
+      width = w.Get(parameters) ?? 0;
+    if (entry.Ints?.TryGetValue(InventoryHeightHash, out var h) == true)
+      height = h.Get(parameters) ?? 0;
+    if (width > 0 && height > 0)
+      return new Vector2i(width, height);
+
+    if (width <= 0)
+      width = zdo.GetInt(InventoryWidthHash, 0);
+    if (height <= 0)
+      height = zdo.GetInt(InventoryHeightHash, 0);
+    if (width > 0 && height > 0)
+      return new Vector2i(width, height);
+
+    // Field might only change one dimension, which complicates the logic.
+    var obj = ZNetScene.instance.GetPrefab(zdo.GetPrefab());
+    var container = obj.GetComponentInChildren<Container>();
+    if (container)
+    {
+      if (width <= 0)
+        width = container.m_width;
+      if (height <= 0)
+        height = container.m_height;
+    }
+    if (width <= 0) width = 4;
+    if (height <= 0) height = 2;
+    return new Vector2i(width, height);
+  }
 }