+++ publish/CHANGELOG.md
@@ -1,6 +1,7 @@
 - v1.43
   - Adds a field `bannedLocationDistance` so that different distance can be used for required and banned locations.
   - Adds support of false value to the poke field `self` to ignore the object itself.
+  - Adds automatic container size detection to simplify adding items.
   - Changes the parameter `<hash_>` to also work for location ids.
   - Fixes quoted strings being split up when used as "type, key, value" in the field `data`.
   - Fixes field value queries not checking child objects (for example Boat chests).